Bind to HH weight branches
Fixes https://github.com/HEP-KBFI/TallinnAnalysis/issues/4. The issue
was not that the input file was missing the HH weights and/or the
LHE-level mHH and cos(theta*) were missing, it was just that the
branch addresses were never bound. The output histograms now give
reasonable results. Removed the lines that were previously added to
demonstrate the issue.